Request-specific configuration for dynamic page components.
Provides the option to supply additional configuration information to components on a per-request basis. This was required to support the use of dynamic HTML on the page and keep a clean separation between between dynamic content and actual form data.

The DynamicHtml component is the first implementation of this approach.