Support custom views.
Rather than just using its built-in view definition, an alternative view file can be specified and the handler will render that instead.

Doing this negates much of the benefit of using this plugin, but it does at least allow reuse of the validation logic and provides a relatively simple fallback for cases where some specific bespoke tweaks are required to a page.